// src/api/projectcontrols/peopleculture/staff-fte/types.ts
export interface O2NL_Staff {
  Discipline_Manager: string;
  Team: string;
  Location: string;
  NOP_Type: string;
  Org: string;
  Project_Role_Title: string;
  Job_Code: string;
  Phase: string;
  Name: string;
  Status: string;
  Last_updated_conf: string;
  //Split_Assignment_: string;
  Resource_Options: string;
  Taitoko_Levin_Site_Based_ : string;
  //Site_Based_: string;
  Pricing_P_G___Prof___Direct_Works: string;
  FTE__AVE_: number;
  Required_Start: Date;
  Required_Finish: Date;
  September_24: number;
  October_24: number;
  November_24: number;
  December_24: number;
  January_25: number;
  February_25: number;
  March_25: number;
  April_25: number;
  May_25: number;
  June_25: number;
  July_25: number;
  August_25: number;
  September_25: number;
  October_25: number;
  November_25: number;
  December_25: number;
  January_26: number;
  February_26: number;
  March_26: number;
  April_26: number;
  May_26: number;
  June_26: number;
  July_26: number;
  August_26: number;
  September_26: number;
  October_26: number;
  November_26: number;
  December_26: number;
  January_27: number;
  February_27: number;
  March_27: number;
  April_27: number;
  May_27: number;
  June_27: number;
  July_27: number;
  August_27: number;
  September_27: number;
  October_27: number;
  November_27: number;
  December_27: number;
  January_28: number;
  February_28: number;
  March_28: number;
  April_28: number;
  May_28: number;
  June_28: number;
  July_28: number;
  August_28: number;
  September_28: number;
  October_28: number;
  November_28: number;
  December_28: number;
  January_29: number;
  February_29: number;
  March_29: number;
  April_29: number;
  May_29: number;
  June_29: number;
  July_29: number;
  August_29: number;
  September_29: number;
  October_29: number;
  November_29: number;
  December_29: number;
  January_30: number;
  February_30: number;
  March_30: number;
  April_30: number;
  May_30: number;
  June_30: number;
  July_30: number;
  August_30: number;
  September_30: number;
  October_30: number;
  November_30: number;
  December_30: number;
}